clean: removing code is related with part which would not use anymore
