+++ core/application/impl/kagome_application_impl.hpp
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef KAGOME_CORE_APPLICATION_IMPL_VALIDATING_NODE_APPLICATION_HPP
-#define KAGOME_CORE_APPLICATION_IMPL_VALIDATING_NODE_APPLICATION_HPP
+#ifndef KAGOME_APPLICATION_ALLINONEAPPLICATION
+#define KAGOME_APPLICATION_ALLINONEAPPLICATION
 
 #include "application/kagome_application.hpp"
 
@@ -15,7 +15,7 @@
 
 namespace kagome::application {
 
-  class ValidatingNodeApplication : public KagomeApplication {
+  class KagomeApplicationImpl : public KagomeApplication {
     template <class T>
     using sptr = std::shared_ptr<T>;
 
@@ -23,15 +23,15 @@ namespace kagome::application {
     using uptr = std::unique_ptr<T>;
 
    public:
-    ~ValidatingNodeApplication() override = default;
+    ~KagomeApplicationImpl() override = default;
 
-    explicit ValidatingNodeApplication(const AppConfiguration &config);
+    explicit KagomeApplicationImpl(const AppConfiguration &config);
 
     void run() override;
 
    private:
     const AppConfiguration &app_config_;
-    uptr<injector::ValidatingNodeInjector> injector_;
+    uptr<injector::KagomeNodeInjector> injector_;
     log::Logger logger_;
 
     sptr<boost::asio::io_context> io_context_;
@@ -49,4 +49,4 @@ namespace kagome::application {
 
 }  // namespace kagome::application
 
-#endif  // KAGOME_CORE_APPLICATION_IMPL_VALIDATING_NODE_APPLICATION_HPP
+#endif  // KAGOME_APPLICATION_ALLINONEAPPLICATION
