+++ core/injector/application_injector.hpp
@@ -45,51 +45,6 @@ namespace kagome {
 namespace kagome::injector {
 
   /**
-   * Dependency injector for a syncing node. Provides all major components
-   * required by the kagome syncing node application.
-   */
-  class SyncingNodeInjector final {
-   public:
-    explicit SyncingNodeInjector(const application::AppConfiguration &);
-
-    std::shared_ptr<application::ChainSpec> injectChainSpec();
-    std::shared_ptr<application::AppStateManager> injectAppStateManager();
-    std::shared_ptr<boost::asio::io_context> injectIoContext();
-    std::shared_ptr<network::Router> injectRouter();
-    std::shared_ptr<network::PeerManager> injectPeerManager();
-    std::shared_ptr<api::ApiService> injectRpcApiService();
-    std::shared_ptr<network::SyncProtocolObserver> injectSyncObserver();
-    std::shared_ptr<consensus::babe::Babe> injectBabe();
-    std::shared_ptr<consensus::grandpa::Grandpa> injectGrandpa();
-
-   protected:
-    std::shared_ptr<class SyncingNodeInjectorImpl> pimpl_;
-  };
-
-  /**
-   * Dependency injector for a validating node. Provides all major components
-   * required by the kagome validating node application.
-   */
-  class ValidatingNodeInjector final {
-   public:
-    explicit ValidatingNodeInjector(const application::AppConfiguration &);
-
-    std::shared_ptr<application::ChainSpec> injectChainSpec();
-    std::shared_ptr<application::AppStateManager> injectAppStateManager();
-    std::shared_ptr<boost::asio::io_context> injectIoContext();
-    std::shared_ptr<network::Router> injectRouter();
-    std::shared_ptr<network::PeerManager> injectPeerManager();
-    std::shared_ptr<api::ApiService> injectRpcApiService();
-    std::shared_ptr<clock::SystemClock> injectSystemClock();
-    std::shared_ptr<network::SyncProtocolObserver> injectSyncObserver();
-    std::shared_ptr<consensus::babe::Babe> injectBabe();
-    std::shared_ptr<consensus::grandpa::Grandpa> injectGrandpa();
-
-   protected:
-    std::shared_ptr<class ValidatingNodeInjectorImpl> pimpl_;
-  };
-
-  /**
    * Dependency injector for a universal node. Provides all major components
    * required by the kagome application.
    */
