--- node/all_in_one/CMakeLists.txt
@@ -1,17 +1,0 @@
-#
-# Copyright Soramitsu Co., Ltd. All Rights Reserved.
-# SPDX-License-Identifier: Apache-2.0
-#
-
-add_executable(kagome_node
-    main.cpp
-    )
-target_link_libraries(kagome_node
-    Boost::boost
-    Boost::program_options
-    kagome_application
-    app_config_impl
-    p2p::p2p_identify
-    p2p::p2p_ping
-    )
-kagome_install(kagome_node)
